smarter walking, more useful plots, samples are saved.
-Likelihood function was optimized for faster logical checks.

-Trim function was made useless by more intelligent initial placement
of walkers. Walkers are now placed within the constrained range, and
are evenly spaced. This is MUCH faster than trimming and means fewer
samples are necessary for a good fit. Walker plots are also more useful
now.

-MCMC function is now fully generalized between 6 or 4 dimensions,
smaller, and much more readable.

-Nearly all functions are now fully documented and commented.

-RV plots are a bit better looking

-Samples from the random walks are now saved to disk, and named based
on the error of the run. The samples produced by a random walk can now
be reanalyzed!