Mcmc and plot tweaks
Added values back into the corner plots, because quantiles are now
being used again to find values. This is because the likelihood
function has also been tweaked, to represent a proper probability
distribution (not just using error as a proxy). This results in much
tighter convergence of walkers, and more Gaussian, independent values.

Phased out use of minimize function in master.py. Too computationally
heavy, and unnecessary when using quantiles to report values.